-added compatibility with chiseledfirepit from mod ChiselTools -the gatherer ability can now benefits all drops instead just the first one -fixed some disease healing related stuff

+++ mods/xlib/src/xeffects/Trigger/AttributeTrigger.cs
@@ -333,7 +333,7 @@ namespace XLib.XEffects
         public override void FromTree(ITreeAttribute tree)
         {
             base.FromTree(tree);
-            RecoveryThreshold = tree.GetFloat("recoverythreshold");
+            RecoveryThreshold = (float)tree.GetDecimal("recoverythreshold");
             ArrayAttribute<TreeAttribute> weightsTree = (tree as TreeAttribute)?.GetAttribute("weights") as ArrayAttribute<TreeAttribute>;
             if (weightsTree == null) return;
             foreach(TreeAttribute attribute in weightsTree.value)
